+++ Tests/BugsnagSessionTrackerStopTest.m
@@ -19,7 +19,7 @@
 
 - (void)setUp {
     [super setUp];
-    self.configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1];
+    self.configuration = [[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1 error:nil];
     self.configuration.autoTrackSessions = NO;
     self.tracker = [[BugsnagSessionTracker alloc] initWithConfig:self.configuration postRecordCallback:nil];
 }